rename mergelines file to merge_lines
