+++ src/symbol/merge_lines.ts
